Fold these types into the same file
There's no need for them to live separately, really.
