+++ src/flickypedia/types/wikimedia.py
@@ -9,17 +9,3 @@ class UserInfo(typing.TypedDict):
 class ShortCaption(typing.TypedDict):
     language: str
     text: str
-
-
-class TitleValidationResult:
-    Ok = typing.TypedDict("Ok", {"result": typing.Literal["ok"]})
-    Failed = typing.TypedDict(
-        "Failed",
-        {
-            "result": typing.Literal["blacklisted", "duplicate", "invalid", "too_long"],
-            "text": str,
-        },
-    )
-
-
-TitleValidation = TitleValidationResult.Ok | TitleValidationResult.Failed
